chore(usemediaquery): convert tests to use RTL
